+++ packages/Chem/src/tests/menu-tests-chem-space.ts
@@ -31,7 +31,8 @@ category('top menu chem space', async () => {
   });
 
   test('chemSpaceOpens.smiles', async () => {
-    const df = DG.Test.isInBenchmark ? await grok.data.files.openTable('Samples:Files/chem/smiles_100K.zip') : smallDf;
+    const df = DG.Test.isInBenchmark ? await grok.data.files
+      .openTable('System:AppData/Chem/tests/smiles_100K.zip') : smallDf;
     await _testChemSpaceReturnsResult(df, 'smiles');
   });